+++ spec/support/models/raw_user.rb
@@ -4,4 +4,6 @@ require 'support/models/application_record'
 
 class RawUser < ApplicationRecord
   self.table_name = "users"
+
+  has_many :authored_books, class_name: "RawBook", foreign_key: :author_id, inverse_of: :author
 end